requirements.txt
Django==5.1.3
djangorestframework==3.15.1
gunicorn==21.2.0
django-cors-headers==4.3.1
whitenoise==6.6.0
Pillow